+++ packages/mobile/App/ui/components/ServerSelectorField/ServerSelector.tsx
@@ -8,9 +8,7 @@ import { StyledText, StyledView } from '../../styled/common';
 import { theme } from '../../styled/theme';
 import { Orientation, screenPercentageToDP } from '../../helpers/screen';
 
-const META_SERVER = __DEV__
-  ? 'https://meta-dev.tamanu.io'
-  : 'https://meta.tamanu.io';
+const META_SERVER = __DEV__ ? 'https://meta-dev.tamanu.io' : 'https://meta.tamanu.io';
 
 type Server = {
   name: string;
@@ -22,12 +20,12 @@ const fetchServers = async (): Promise<SelectOption[]> => {
   // To use a local server, just edit this and select it.
   // The sync server config is sticky, so you can safely revert it after
   // the first sync begins and it'll stay connecting to your local server.
-  // return [{ label: 'Local', value: 'http://192.168.0.1:3000' }];
+  return [{ label: 'Local', value: 'http://192.168.1.124:3000' }];
 
   const response = await fetch(`${META_SERVER}/servers`);
   const servers: Server[] = await response.json();
 
-  return servers.map((s) => ({ label: s.name, value: s.host }));
+  return servers.map(s => ({ label: s.name, value: s.host }));
 };
 
 export const ServerSelector = ({ onChange, label, value }): ReactElement => {
@@ -45,25 +43,21 @@ export const ServerSelector = ({ onChange, label, value }): ReactElement => {
     })();
   }, [netInfo.isInternetReachable]);
 
-  const onServerSelected = useCallback((server) => {
-    setDisplayValue(server ? server.label : '');
-    onChange(server?.value);
-  }, [onChange]);
+  const onServerSelected = useCallback(
+    server => {
+      setDisplayValue(server ? server.label : '');
+      onChange(server?.value);
+    },
+    [onChange],
+  );
 
   if (!netInfo.isInternetReachable) {
-    return (
-      <StyledText color={theme.colors.ALERT}>
-        No internet connection available.
-      </StyledText>
-    );
+    return <StyledText color={theme.colors.ALERT}>No internet connection available.</StyledText>;
   }
 
   return (
     <>
-      <StyledView
-        marginBottom={10}
-        height={screenPercentageToDP(4.86, Orientation.Height)}
-      >
+      <StyledView marginBottom={10} height={screenPercentageToDP(4.86, Orientation.Height)}>
         <InputContainer>
           <StyledText
             color={theme.colors.TEXT_DARK}
